move SPDateCreated to its own file and implement its specs
