+++ ConsentViewController/Classes/SourcePointClient/ConsentStatusResponse.swift
@@ -7,28 +7,6 @@
 
 import Foundation
 
-struct SPDateCreated: Codable, Equatable {
-    static let format: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"
-        return formatter
-    }()
-
-    let originalDateString: String
-    let date: Date
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        originalDateString = try container.decode(String.self)
-        date = SPDateCreated.format.date(from: originalDateString)!
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        try container.encode(originalDateString)
-    }
-}
-
 struct ConsentStatusResponse: Decodable, Equatable {
     struct Data: Decodable, Equatable {
         struct GDPR: Decodable, Equatable {
